+++ packages/tslint-config-angular/index.js
@@ -1,7 +1,4 @@
 module.exports = {
-  extends: [
-    '@codemakebros/tslint-config',
-    'tslint-angular'
-  ],
+  extends: ['@codemakebros/tslint-config', 'tslint-angular'],
   rules: {}
 };