+++ packages/stylelint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for @codemakebros/stylelint-config
 
+## 0.0.2-rc4 (Apr 12, 2018)
+
+* Done except `[Stylistic issues]Media..., At-rule...`
+
 ## 0.0.2-rc3 (Apr 11, 2018)
 
 * Add rules until `[Stylistic issues]declaration-block-trailing-semicolon`