+++ packages/tslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for @codemakebros/tslint-config
 
+## 0.0.2-rc4 (Apr 12, 2018)
+
+* Complete all `tslint-microsoft-contrib` rules
+* Add `allow-empty-class` option into `no-unnecessary-class` rule
+* Deactivate `[microsoft]missing-jsdoc` rule
+
 ## 0.0.2-rc3 (Apr 11, 2018)
 
 * Complete all `tslint` rules